Fix scheme names and make private function names consistent
For the private function names I've opted to make them consistent with
the scheme names, since those are presumably the most natural ways to
refer to them.

Plan is to make the private functions public, and document those
functions with full details. Eventually we can probably delete the
new_predefined_driven_control, although it's fine to stick around for a
while for compatibility purposes.

https://q-ctrl.atlassian.net/browse/QENG-1118