Refactor a parameter-checking function for predefined driven-control functions (#146)
This is the first PR of another series of PRs for refactoring the driven control module, which will include:
- remove the wrapper function `new_predefined_driven_control`
- remove unnecessary inputs, if any
- remove unnecessary default values
- use check function rather than raising errors
- format tweak, if necessary

As the first step, this PR refactors a private checking function shared by some predefined driven control functions. The `float` type cast is removed and now it would only do the checking, but not returning any value. The name is also tweaked, to better reflect its functionality.