Update driven control notebook (#151)
In preparation of removing `new_primitive_control`, here we use specicifc functions to create driven controls.

Note that the the values of `amplitude_x` , `amplitude_y` are different from old values. Looking at the history,  the reason is we changed how these values are calculated in #115. The notebooks were added in  #71, but not updated after merging #115.