feat: Add function to resample a driven control (#201)
Alternative to #200. Even though this one is a bit less consistent with BOULDER OPAL, I think it's better. For a start, it means the user can also benefit from resampling for the export-to-file function too, which I would have thought should be very useful. Also, I don't think consistency with BOULDER OPAL is that important here, because DrivenControl is a very different beast to a plain old pulse.